+++ db/migrate/20191220011006_add_locked_to_transactions.rb
@@ -1,0 +1,5 @@
+class AddLockedToTransactions < ActiveRecord::Migration[6.0]
+  def change
+    add_column :transactions, :locked, :boolean, default: false
+  end
+end