+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] changes
 
+## [v1.6.0] - 2020-02-05
+
+## Added
+- Account Stash feature added to assist in putting money aside and out of site for planned expenses.
+- Transaction locking feature was implemented and enabled for starting balance transactions as well as stash-related transactions
+
+## Changed
+- Bulma version bumped to 0.8.0
+- Ruby version bumped to 2.6.5
+- Bundler version bumped to 2.0.2
+- will_paginate gem swapped for pagy gem
+- Transactions screen now displays total Pending transactions and total Stashed balance
+
 ## [v1.5.0] - 2019-12-19
 
 ### Changed
@@ -147,7 +160,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Account overview with list of all accounts and balances for each
 
 
-[Unreleased]: https://github.com/odinsride/olubalance/compare/v1.5.0...HEAD
+[Unreleased]: https://github.com/odinsride/olubalance/compare/v1.6.0...HEAD
+[v1.6.0]: https://github.com/odinsride/olubalance/compare/v1.5.0...v1.6.0
 [v1.5.0]: https://github.com/odinsride/olubalance/compare/v1.4.5...v1.5.0
 [v1.4.5]: https://github.com/odinsride/olubalance/compare/v1.4.2...v1.4.5
 [v1.4.2]: https://github.com/odinsride/olubalance/compare/v1.4.1...v1.4.2